Add extended-define-syntax
This provides (define-syntax (f x) ...) as a shorthand for
(define-syntax (lambda (x) ...)). This works out of the box in Chez, but
not Guile. In Racket it works normally, but not in R6RS mode.

This allows eliminating a level of indentation in the SICP macro
definition (previously just indented incorrectly). It requires importing
(for (rnrs base (6)) run expand) in racket.ss so that the use of lambda
in the extended-define-syntax definition works. See:

https://www.ccs.neu.edu/home/ryanc/macro-patterns/error_unbound-in-transformer-env.html