Export register functions instead of directly registering entries
This changes the SICP macro to look like `(SICP register e* ...)` and
expand to `(define (register) ...)`, allowing main.ss to control the
order of registration. Previously a fake dependency chain (chapter 5
depends on chapter 4 depends on ... chapter 1) was used to ensure the
default order in `./run.sh chez` was correct. This is now removed, and
instead main.ss just registers in the correct order.

The main benefit of this is speeding up Racket compile times. Before,
any change to chapter N would cause recompilation of all chapters from N
to 5. Now, it only causes recompilation of that individual chapter.