Use guile auto-compilation
This changes run.sh to use Guile's auto-compilation, but -O1 instead of
the default of -O2 because that takes 10x longer. I'm not sure why I
wanted to be able to mutate quoted lists, when it's only a couple places
that were relying on it, and it seems outside the R6RS spec anyway.

Now guile is competitive with chez for time!