Highlight quoted things, and keep arrows for console
For example 'foo and '(foo bar) now get quoted yellow/brown, whereas
strings and numbers are still green. I made scheme.xml support up to 15
levels of nesting. And now the console output is in grey, just
light/dark enough to get 4.5 contrast (keeping the arrows like => below
that contrast since they are more decoration, not to be "read") and
italic, with the Unicode arrow which looks nice.