Nest code blocks in (a), (b), etc. parts
This improves the markup so that there is really a single <ol> for
exercises with multiple parts, rather than indivual ones with later ones
having start="2" for example. It looks the same because I dedent the
code blocks (and diagrams). But now the markup is clear: to continue
within the (a) after the code, keep it indented.